% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\docType{data}
\name{definitions}
\alias{definitions}
\title{definitions: Definitions of terms in the WHO Glossary (https://www.who.int/publications/i/item/9789240105485)}
\format{
A tibble with 207 rows and 2 variables
\describe{
  \item{Term}{Term in the Glossary}
  \item{Description}{Description of the term}
}
}
\usage{
definitions
}
\description{
This datasets contains all terms and their definitions as defined in the WHO Glossary.
}
\keyword{datasets}
