% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\docType{data}
\name{themes}
\alias{themes}
\title{themes: Table of Thematic Areas from https://www.who.int/publications/i/item/9789240105485}
\format{
A tibble with 217 rows and 2 variables
\describe{
  \item{Thematic Area}{Thematic Area}
  \item{Term}{Term in the Glossary}
}
}
\usage{
themes
}
\description{
A table that assigns thematic areas to definitions included in the WHO Glossary (https://www.who.int/publications/i/item/9789240105485)
}
\keyword{datasets}
